# Reuse handle when bulk-reloading