@noctux 2018-02-20
> s/0/nullptr/ 

Urgs, yeah, of course. I don't know how that got there... fixed now. By the way, did I mention that styleguide thing? :)